% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{bootsens.os}
\alias{bootsens.os}
\alias{extrema.os}
\title{Sensitivity analysis for observational studies}
\usage{
bootsens.os(A, X, Y, gamma = 0, alpha = 0.05, estimand = c("ate", "att"),
  reg.adjust = FALSE, parallel = TRUE, B = 1000, warm.start = FALSE)

extrema.os(A, X, Y, gamma = 0, estimand = c("ate", "att"),
  reg.adjust = FALSE, start = NULL)
}
\arguments{
\item{A}{Indicator of missingness}

\item{X}{A matrix of covariates}

\item{Y}{Outcome}

\item{gamma}{Sensitivity parameter (log odds ratio)}

\item{alpha}{Significance level}

\item{estimand}{Either "ate" (average treatment effect) or "att" (average treatment effect on the treated)}

\item{reg.adjust}{Should regression adjustment (augmented IPW) be used?}

\item{parallel}{Should parallel computing be used?}

\item{B}{Number of Bootstrap resamples.}

\item{warm.start}{Warm start the refitting of propensity score model (doesn't seem to help).}

\item{start}{Starting values for the propensity score model to warm start \code{glm}.}
}
\value{
A (1 - alpha) confidence interval.

Extrema (an interval).
}
\description{
Sensitivity analysis for observational studies
}
\section{Functions}{
\itemize{
\item \code{extrema.os}: Obtain extrema of IPW estimator for observational studies
}}

\examples{

\donttest{
## Fish consumption and blood mercury dataset
require(CrossScreening)
data(nhanes.fish)
A <- as.numeric(nhanes.fish$fish.level == "high")
X <- nhanes.fish[, c("gender", "age", "income", "income.missing",
"race", "education", "smoking.ever", "smoking.now")]
X$race <- factor(X$race)
X1 <- model.matrix(~ . - 1, X)
Y <- log2(nhanes.fish$o.LBXTHG)

## Assuming no unmeasure confounder (i.e. gamma = 0)
extrema.os(A, X1, Y) # point estimate
bootsens.os(A, X1, Y) # confidence interval

## Sensitivity analysis (gamma = 1)
extrema.os(A, X1, Y, gamma = 1) # point estimate
bootsens.os(A, X1, Y, gamma = 1) # confidence interval
}

}
