% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{extrema.md}
\alias{extrema.md}
\title{Obtain extrema of IPW estimator for missing data}
\usage{
extrema.md(A, X, Y, gamma = 0, estimand = c("all", "missing"),
  reg.adjust = FALSE, start = NULL)
}
\arguments{
\item{A}{Indicator of missingness}

\item{X}{A matrix of covariates}

\item{Y}{Outcome}

\item{gamma}{Sensitivity parameter (log odds ratio)}

\item{estimand}{Target estimand, either E[Y] ("all") or E[Y|A=0] ("missing")}

\item{reg.adjust}{Should regression adjustment (augmented IPW) be used?}

\item{start}{Starting values for the propensity score model to warm start \code{glm}.}
}
\value{
Extrema (an interval).
}
\description{
Obtain extrema of IPW estimator for missing data
}
