% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{extrema.os}
\alias{extrema.os}
\title{Obtain extrema of IPW estimator for observational studies}
\usage{
extrema.os(A, X, Y, gamma = 0, estimand = c("ate", "att"),
  reg.adjust = FALSE, start = NULL)
}
\arguments{
\item{A}{Indicator of missingness}

\item{X}{A matrix of covariates}

\item{Y}{Outcome}

\item{gamma}{Sensitivity parameter (log odds ratio)}

\item{estimand}{Either "ate" (average treatment effect) or "att" (average treatment effect on the treated)}

\item{reg.adjust}{Should regression adjustment (augmented IPW) be used?}

\item{start}{Starting values for the propensity score model to warm start \code{glm}.}
}
\value{
Extrema (an interval).
}
\description{
Obtain extrema of IPW estimator for observational studies
}
