% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{get.extrema}
\alias{get.extrema}
\title{Fast solution to obtain extrema of IPW estimator in missing data
problems}
\usage{
get.extrema(A, X, Y, gamma = 0, fitted.prob, estimand = c("all", "missing"))
}
\arguments{
\item{A}{Indicator of missingness}

\item{X}{A matrix of covariates}

\item{Y}{Outcome}

\item{gamma}{Sensitivity parameter (log odds ratio)}

\item{fitted.prob}{Fitted propensity score}

\item{estimand}{Target estimand, either E[Y] ("all") or E[Y|A=0] ("missing")}

\item{discrete}{Are the data (both X and Y) discrete?}
}
\value{
Extrema (an interval) of the IPW estimator.
}
\description{
Fast solution to obtain extrema of IPW estimator in missing data
problems
}
