% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{bootsens.md}
\alias{bootsens.md}
\alias{extrema.md}
\title{Sensitivity analysis for missing data}
\usage{
bootsens.md(
  A,
  X,
  Y,
  gamma = 0,
  alpha = 0.05,
  estimand = c("all", "missing"),
  reg.adjust = FALSE,
  parallel = TRUE,
  B = 1000,
  warm.start = FALSE
)

extrema.md(
  A,
  X,
  Y,
  gamma = 0,
  estimand = c("all", "missing"),
  reg.adjust = FALSE,
  start = NULL
)
}
\arguments{
\item{A}{Indicator of missingness}

\item{X}{A matrix of covariates}

\item{Y}{Outcome}

\item{gamma}{Sensitivity parameter (log odds ratio)}

\item{alpha}{Significance level}

\item{estimand}{Target estimand, either E[Y] ("all") or E[Y|A=0] ("missing")}

\item{reg.adjust}{Should regression adjustment (augmented IPW) be used?}

\item{parallel}{Should parallel computing be used?}

\item{B}{Number of Bootstrap resamples.}

\item{warm.start}{Warm start the refitting of propensity score model (doesn't seem to help).}

\item{start}{Starting values for the propensity score model to warm start \code{glm}.}
}
\value{
A (1 - alpha) confidence interval.

Extrema (an interval).
}
\description{
Sensitivity analysis for missing data
}
\section{Functions}{
\itemize{
\item \code{extrema.md}: Obtain extrema of IPW estimator for missing data
}}

